temp file service trait
