+++ app/files/WkTempFileService.scala
@@ -1,70 +1,13 @@
 package files
 
-import cleanup.CleanUpService
-import com.scalableminds.util.time.Instant
-
-import java.nio.file.{Files, Path, Paths}
-import com.scalableminds.util.tools.Fox
-import com.typesafe.scalalogging.LazyLogging
+import com.scalableminds.webknossos.tracingstore.cleanup.CleanUpService
+import com.scalableminds.webknossos.tracingstore.files.TempFileService
 
 import javax.inject.Inject
-import net.liftweb.common.Box.tryo
-import org.apache.commons.io.FileUtils
 
 import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.{DurationInt, FiniteDuration}
-import scala.util.Random
-
-/**
-  * Avoiding Java TemporaryFiles because of seeming openJDK regression,
-  * see discussion at https://github.com/scalableminds/webknossos/issues/6173
-  */
-trait TempFileService extends LazyLogging {
-  protected def cleanUpService: CleanUpService
-  implicit protected def ec: ExecutionContext
-  protected def prefix: String
-
-  private val tmpDir: Path = Paths.get(System.getProperty("java.io.tmpdir")).resolve(s"${prefix}TempFiles")
-
-  private val activeTempFiles = scala.collection.mutable.Set[(Path, Instant)]()
-
-  cleanUpService.register("Clean up expired temporary files", 1 hour)(cleanUpExpiredFiles())
-
-  private def ensureParent(): Path =
-    Files.createDirectories(tmpDir)
-
-  def create(prefix: String = "tmpFile", lifeTime: FiniteDuration = 2 hours): Path = {
-    ensureParent()
-    val path = tmpDir.resolve(f"$prefix-${Random.alphanumeric.take(15).mkString("")}")
-    logger.debug(f"Creating temp file at $path")
-    Files.createFile(path)
-    activeTempFiles.add((path, Instant.now + lifeTime))
-    path
-  }
-
-  private def cleanUpExpiredFiles(): Fox[Unit] = {
-    val now = Instant.now
-    activeTempFiles.foreach {
-      case (path, expiryTime) =>
-        if (expiryTime < now) {
-          tryo(Files.delete(path))
-          activeTempFiles.remove((path, expiryTime))
-        }
-    }
-    Fox.successful(())
-  }
-
-  def cleanUpAll(): Unit =
-    FileUtils.deleteDirectory(tmpDir.toFile)
-
-}
 
 class WkTempFileService @Inject()(val cleanUpService: CleanUpService)(implicit val ec: ExecutionContext)
     extends TempFileService {
-  override val prefix = "webknossos"
-}
-
-class TsTempFileService @Inject()(val cleanUpService: CleanUpService)(implicit val ec: ExecutionContext)
-    extends TempFileService {
-  override val prefix = "webknossosTracingStore"
+  override val moduleName = "webknossos"
 }
